EdgeYoloDepth/requirements.txt
# Train model 
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.3.1
idna==3.4
Jinja2==3.1.3
kiwisolver==1.4.5
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.2
pathlib==1.0.1
pillow==10.2.0
protobuf==5.28.3
pycocotools==2.0.7
Pygments==2.15.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
tabulate==0.9.0
thop==0.1.1.post2209072238
tqdm==4.65.0
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.1.0

# Eval model (for depth)
statsmodels==0.14.4
openpyxl==3.1.5

# Convert to ONNX
onnx==1.16.0
onnxsim==0.4.36
onnxruntime==1.20.0
coloredlogs==15.0.1
flatbuffers==24.3.25


